+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - add: support for parquet files and Delta/Hive/Iceberg directories
 - add: support for SAS, SPSS and Stata files
 - add: mssql db support
+- add: gpkg support (sqlite based)
 
 ## 0.2.1 (2026-01-21)